+++ grid-cli/cmd/update_kubernetes.go
@@ -1,0 +1,16 @@
+// Package cmd for parsing command line arguments
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// updateKubernetesCmd represents the update kubernetes command
+var updateKubernetesCmd = &cobra.Command{
+	Use:   "kubernetes",
+	Short: "update kubernetes workers",
+}
+
+func init() {
+	updateCmd.AddCommand(updateKubernetesCmd)
+}