+++ grid-proxy/internal/indexer/health.go
@@ -39,14 +39,30 @@ func (w *HealthWork) Get(ctx context.Context, rmb *peer.RpcClient, twinId uint32
 }
 
 func (w *HealthWork) Upsert(ctx context.Context, db db.Database, batch []types.HealthReport) error {
+	// to prevent having multiple data for the same twin from different finders
+	batch = removeDuplicates(batch)
 	return db.UpsertNodeHealth(ctx, batch)
 }
 
 func getHealthReport(response diagnostics.Diagnostics, twinId uint32) types.HealthReport {
 	report := types.HealthReport{
 		NodeTwinId: twinId,
-		Healthy:    response.Healthy,
+		Healthy:    false,
+		UpdatedAt:  time.Now().Unix(),
 	}
 
 	return report
+}
+
+func removeDuplicates(reports []types.HealthReport) []types.HealthReport {
+	seen := make(map[uint32]bool)
+	result := []types.HealthReport{}
+	for _, report := range reports {
+		if _, ok := seen[report.NodeTwinId]; !ok {
+			seen[report.NodeTwinId] = true
+			result = append(result, report)
+		}
+	}
+
+	return result
 }