+++ tfrobot/pkg/deployer/types.go
@@ -36,12 +36,14 @@ type Vms struct {
 	SSDDisks   []Disk            `yaml:"ssd" json:"ssd"`
 	PublicIP4  bool              `yaml:"public_ip4" json:"public_ip4"`
 	PublicIP6  bool              `yaml:"public_ip6" json:"public_ip6"`
-	Planetary  bool              `yaml:"planetary" json:"planetary"`
+	Ygg        bool              `yaml:"ygg_ip" json:"ygg_ip"`
+	Mycelium   bool              `yaml:"mycelium_ip" json:"mycelium_ip"`
 	Flist      string            `yaml:"flist" validate:"required" json:"flist"`
 	RootSize   uint64            `yaml:"root_size" validate:"max=10240" json:"root_size"` // max 10 TB
 	Entrypoint string            `yaml:"entry_point" validate:"required" json:"entry_point"`
 	SSHKey     string            `yaml:"ssh_key" validate:"required" json:"ssh_key"`
 	EnvVars    map[string]string `yaml:"env_vars" json:"env_vars"`
+	WireGuard  bool              `yaml:"wireguard" json:"wireguard"`
 }
 
 type Disk struct {
@@ -59,7 +61,8 @@ type vmOutput struct {
 	NetworkName string
 	PublicIP4   string
 	PublicIP6   string
-	PlanetaryIP string
+	YggIP       string
+	MyceliumIP  string
 	IP          string
 	Mounts      []workloads.Mount
 	NodeID      uint32